pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fogcutter"
description = "Cutting through LLM uncertainty with modern metrics"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Yathish Sadashivareddy", email = "Yathishsadashivareddy@gmail.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]

dependencies = [
  "torch>=2.0.0",
  "transformers>=4.30.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "pandas>=2.3.3",
  "pytest>=8.4.2",
  "ruff>=0.14.10",
  "build>=1.3.0",
  "openai>=1.0.0",
  "google-genai>=1.0.0",
  "sentence-transformers>=5.1.2",
  "python-dotenv>=1.2.1",
  "pydantic-settings>=2.11.0",
  "vertexai>=1.71.1",
  "google-generativeai>=0.8.6",
  "google-cloud-aiplatform>=1.71.1",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/fogcutter"
Repository = "https://github.com/YOUR_USERNAME/fogcutter"

[tool.hatch.version]
path = "src/fogcutter/__init__.py"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
